// src/utils/printer/types.ts
export interface TicketData {
  ticket: string;
  date: string;
  employee: string;
  items: SaleItem[];
  subtotal: number;
  tax: number;
  discount: number;
  total: number;
  received: number;
  change: number;
  method: string;
  width: number;
}